pyproject.toml
[project]
name = "bdo-pvp-analytics-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
